move request_builder into internal pkg (#304)
