--- request_builder.go
@@ -1,42 +1,0 @@
-package openai
-
-import (
-	"bytes"
-	"context"
-	"net/http"
-
-	utils "github.com/sashabaranov/go-openai/internal"
-)
-
-type requestBuilder interface {
-	build(ctx context.Context, method, url string, request any) (*http.Request, error)
-}
-
-type httpRequestBuilder struct {
-	marshaller utils.Marshaller
-}
-
-func newRequestBuilder() *httpRequestBuilder {
-	return &httpRequestBuilder{
-		marshaller: &utils.JSONMarshaller{},
-	}
-}
-
-func (b *httpRequestBuilder) build(ctx context.Context, method, url string, request any) (*http.Request, error) {
-	if request == nil {
-		return http.NewRequestWithContext(ctx, method, url, nil)
-	}
-
-	var reqBytes []byte
-	reqBytes, err := b.marshaller.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-
-	return http.NewRequestWithContext(
-		ctx,
-		method,
-		url,
-		bytes.NewBuffer(reqBytes),
-	)
-}
